Switch to `ListenToLedgerUpdates` stream
